@@ -18,45 +18,45 @@
   "private": true,
   "dependencies": {
     "@agm/core": "1.0.0-beta.5",
-    "@angular/animations": "7.2.4",
-    "@angular/cdk": "7.3.1",
-    "@angular/common": "7.2.4",
-    "@angular/compiler": "7.2.4",
-    "@angular/core": "7.2.4",
-    "@angular/elements": "7.2.4",
-    "@angular/forms": "7.2.4",
-    "@angular/material": "7.3.1",
-    "@angular/platform-browser": "7.2.4",
-    "@angular/platform-browser-dynamic": "7.2.4",
-    "@angular/router": "7.2.4",
+    "@angular/animations": "8.0.0-beta.9",
+    "@angular/cdk": "7.3.5",
+    "@angular/common": "8.0.0-beta.9",
+    "@angular/compiler": "8.0.0-beta.9",
+    "@angular/core": "8.0.0-beta.9",
+    "@angular/elements": "8.0.0-beta.9",
+    "@angular/forms": "8.0.0-beta.9",
+    "@angular/material": "7.3.5",
+    "@angular/platform-browser": "8.0.0-beta.9",
+    "@angular/platform-browser-dynamic": "8.0.0-beta.9",
+    "@angular/router": "8.0.0-beta.9",
     "@fortawesome/angular-fontawesome": "0.3.0",
-    "@fortawesome/fontawesome-svg-core": "1.2.14",
-    "@fortawesome/free-brands-svg-icons": "5.7.1",
-    "@fortawesome/free-regular-svg-icons": "5.7.1",
-    "@fortawesome/free-solid-svg-icons": "5.7.1",
-    "@ng-bootstrap/ng-bootstrap": "4.0.2",
-    "@nebular/auth": "3.2.1",
-    "@nebular/security": "3.2.1",
-    "@nebular/theme": "3.2.1",
+    "@fortawesome/fontawesome-svg-core": "1.2.17",
+    "@fortawesome/free-brands-svg-icons": "5.8.1",
+    "@fortawesome/free-regular-svg-icons": "5.8.1",
+    "@fortawesome/free-solid-svg-icons": "5.8.1",
+    "@ng-bootstrap/ng-bootstrap": "4.1.0",
+    "@nebular/auth": "3.4.1",
+    "@nebular/security": "3.4.1",
+    "@nebular/theme": "3.4.1",
     "@webcomponents/custom-elements": "1.2.1",
     "angularx-social-login": "1.2.6",
-    "bootstrap": "4.2.1",
-    "core-js": "2.6.4",
+    "bootstrap": "4.3.1",
+    "core-js": "2.6.5",
     "document-register-element": "1.13.1",
     "hammerjs": "2.0.8",
-    "ngx-infinite-scroll": "7.0.1",
+    "ngx-infinite-scroll": "7.1.0",
     "rxjs": "6.4.0",
-    "tslib": "1.9.0",
+    "tslib": "1.9.3",
     "zone.js": "0.8.29"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.13.1",
-    "@angular-devkit/build-ng-packagr": "0.13.1",
-    "@angular/cli": "7.3.1",
-    "@angular/compiler-cli": "7.2.4",
+    "@angular-devkit/build-angular": "0.800.0-beta.8",
+    "@angular-devkit/build-ng-packagr": "0.800.0-beta.8",
+    "@angular/cli": "8.0.0-beta.8",
+    "@angular/compiler-cli": "8.0.0-beta.9",
     "@types/gapi": "0.0.36",
-    "@types/gapi.auth2": "0.0.49",
-    "@angular/language-service": "7.2.4",
+    "@types/gapi.auth2": "0.0.50",
+    "@angular/language-service": "8.0.0-beta.9",
     "@types/jasmine": "2.8.12",
     "@types/jasminewd2": "2.0.6",
     "@types/node": "10.12.24",
@@ -75,7 +75,7 @@
     "tsickle": "0.34.3",
     "tslib": "1.9.0",
     "tslint": "5.12.1",
-    "typescript": "3.2.4",
+    "typescript": "3.3.4000",
     "webpack-bundle-analyzer": "3.0.3"
   }
 }
