@@ -1,6 +1,5 @@
 import {Pipe, PipeTransform} from '@angular/core';
-import {DomSanitizer} from '@angular/platform-browser';
-import {SafeHtml} from '@angular/platform-browser/src/security/dom_sanitization_service';
+import {DomSanitizer, SafeHtml} from '@angular/platform-browser';
 
 @Pipe({name: 'safeHtml'})
 export class SafeHtmlPipe implements PipeTransform {
