@@ -27,8 +27,7 @@ import {FontAwesomeModule} from '@fortawesome/angular-fontawesome';
 import {ContentLoadingComponent} from './components/content-loading/content-loading.component';
 import {AutoScrollDirective} from './directives/auto-scroll.directive';
 import {ModelComponent} from './components/model/model.component';
-import {createCustomElement} from '@angular/elements';
-import {NgElementConstructor} from '@angular/elements/src/create-custom-element';
+import {createCustomElement, NgElementConstructor} from '@angular/elements';
 import {CfsInfiniteScrollService} from './services/cfs-infinite-scroll.service';
 import {ReadMoreComponent} from './components/read-more/read-more.component';
 import {SafeHtmlPipe} from './pipes/safe-html.pipe';
