@@ -1,7 +1,5 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {AuthModule} from '@candiman/auth';
-import {WINDOW, windowFactory} from "../../../projects/candiman/auth/src/lib/auth.module";
 import {NbAlertModule, NbButtonModule, NbCheckboxModule, NbInputModule, NbLayoutModule} from "@nebular/theme";
 import {AuthRoutingModule} from "../../../projects/candiman/auth/src/lib/auth-routing.module";
 
@@ -10,18 +8,14 @@ import {AuthRoutingModule} from "../../../projects/candiman/auth/src/lib/auth-ro
   ],
   imports: [
     CommonModule,
-    NbLayoutModule,
+    // NbLayoutModule,
     NbAlertModule,
     NbInputModule,
     NbButtonModule,
     NbCheckboxModule,
-    AuthModule
   ],
   providers: [
-    {
-      provide: WINDOW,
-      useFactory: windowFactory
-    }
+
   ]
 })
 export class CfsAuthModule {
